Moved payload pre module to own file
