+++ test/payloadprep.rb
@@ -1,0 +1,62 @@
+module PayloadPrep
+
+  def setup_model_testing_environment
+    TrafficSpy::RegistrationParser.new({"rootUrl"=>"http://jumpstartlab.com", "identifier"=>"jumpstartlab"}).parsing_validating
+
+    TrafficSpy::PayloadParser.new(payload_params1).payload_response
+    TrafficSpy::PayloadParser.new(payload_params2).payload_response
+    TrafficSpy::PayloadParser.new(payload_params3).payload_response
+  end
+
+  def register_user_thru_controller
+    post '/sources', {rootUrl: 'http://jumpstartlab.com', identifier: "jumpstartlab"}
+  end
+
+  def register_user
+    parser_setup({"rootUrl"=>"http://jumpstartlab.com", "identifier"=>"jumpstartlab"}).parsing_validating
+  end
+
+  def parser_setup(params)
+    @parser = TrafficSpy::RegistrationParser.new(params)
+  end
+
+  def payload_params1
+    {"payload"=>
+      "{\"url\":\"http://jumpstartlab.com/blog\",\"requestedAt\":\"2013-02-16 21:38:28 -0700\",\"respondedIn\":39,\"referredBy\":\"http://jumpstartlab.com\",\"requestType\":\"GET\",\"parameters\":[],\"eventName\":\"socialLogin\",\"userAgent\":\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17\",\"resolutionWidth\":\"1920\",\"resolutionHeight\":\"1280\",\"ip\":\"63.29.38.211\"}",
+     "splat"=>[],
+     "captures"=>["jumpstartlab"],
+     "id"=>"jumpstartlab"}
+  end
+
+  def payload_params2
+    {"payload"=>
+      "{\"url\":\"http://jumpstartlab.com/blog\",\"requestedAt\":\"2013-02-16 21:39:28 -0700\",\"respondedIn\":37,\"referredBy\":\"http://jumpstartlab.com\",\"requestType\":\"POST\",\"parameters\":[],\"eventName\":\"socialLogin\",\"userAgent\":\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17\",\"resolutionWidth\":\"1920\",\"resolutionHeight\":\"1280\",\"ip\":\"63.29.38.211\"}",
+     "splat"=>[],
+     "captures"=>["jumpstartlab"],
+     "id"=>"jumpstartlab"}
+  end
+
+  def payload_params3
+    {"payload"=>
+      "{\"url\":\"http://jumpstartlab.com/weather\",\"requestedAt\":\"2013-02-16 19:38:28 -0700\",\"respondedIn\":39,\"referredBy\":\"http://google.com\",\"requestType\":\"POST\",\"parameters\":[1],\"eventName\":\"socialLogin\",\"userAgent\":\"Mozilla/5.0 (Windows-RSS-Platform/2.0 (IE 11.0; Windows NT 6.1)) AppleWebKit/537.75.14 (KHTML, like Gecko) Version/7.0.3 Safari/7046A194A\",\"resolutionWidth\":\"12\",\"resolutionHeight\":\"120\",\"ip\":\"63.29.38.210\"}",
+     "splat"=>[],
+     "captures"=>["jumpstartlab"],
+     "id"=>"jumpstartlab"}
+  end
+
+  def payload_params4
+    {"payload"=>
+      "{\"url\":\"http://jumpstartlab.com/blog\",\"requestedAt\":\"2013-02-16 20:38:28 -0700\",\"respondedIn\":39,\"referredBy\":\"http://cnn.com\",\"requestType\":\"PUT\",\"parameters\":[1],\"eventName\":\"registrationInformation\",\"userAgent\":\"Mozilla/5.0 (Windows-RSS-Platform/2.0 (IE 11.0; Windows NT 6.1)) AppleWebKit/537.75.14 (KHTML, like Gecko) Version/7.0.3 Safari/7046A194A\",\"resolutionWidth\":\"12\",\"resolutionHeight\":\"120\",\"ip\":\"63.29.38.210\"}",
+     "splat"=>[],
+     "captures"=>["jumpstartlab"],
+     "id"=>"jumpstartlab"}
+  end
+
+  def payload_params5
+    {"payload"=>
+      "{\"url\":\"http://jumpstartlab.com/blog\",\"requestedAt\":\"2013-02-16 20:38:28 -0700\",\"respondedIn\":39,\"referredBy\":\"http://reddit.com\",\"requestType\":\"POST\",\"parameters\":[1],\"eventName\":\"registrationInformation\",\"userAgent\":\"Mozilla/5.0 (Windows-RSS-Platform/2.0 (IE 11.0; Windows NT 6.1)) AppleWebKit/537.75.14 (KHTML, like Gecko) Version/7.0.3 Safari/7046A194A\",\"resolutionWidth\":\"12\",\"resolutionHeight\":\"120\",\"ip\":\"63.29.38.210\"}",
+     "splat"=>[],
+     "captures"=>["jumpstartlab"],
+     "id"=>"jumpstartlab"}
+  end
+end
